refactor(core/plugins): move plugin-related util functions
